Fixed Esc not closing the buy menu when using the 1.1 or ioquake engines.
And pressing the buy key a second time won't close the buy menu anymore. This doesn't work in the 1.1 engine anyway.  It was a workaround for a bug where you would sometimes be unable to close the buy menu in any other way.  Hopefully the changed Esc handling will take care of that problem.


git-svn-id: http://svn.code.sf.net/p/smokinguns/code/branches/1.1@269 07eb42f0-b032-4355-afd3-42b1f3832223